Add add-remove feature
